+++ types/protocol-mapping.d.ts
@@ -3046,6 +3046,14 @@ export namespace ProtocolMapping {
             returnType: Protocol.DOM.ForceShowPopoverResponse;
         };
         /**
+         * Returns candidate nodes that are configured as triggers for the given popover.
+         * @experimental
+         */
+        'DOM.getImplicitAnchorCandidates': {
+            paramsType: [Protocol.DOM.GetImplicitAnchorCandidatesRequest];
+            returnType: Protocol.DOM.GetImplicitAnchorCandidatesResponse;
+        };
+        /**
          * When enabling, this API forces an element to gain interest in its target,
          * keeping interest active until disabled.
          * @experimental